Separate re-asking the LLM from deciding what language a text is in
language_check held three subjects and read like it: turning text into a verdict,
describing that verdict, and running a retry loop around an LLM. The third is the
odd one — it is the only part that knows an LLM exists at all, and it is prompt
engineering and control flow rather than detection.

generate_in_language, correction_note and LanguageMismatchError move to a sibling
module of their own. What is left is a detector you can hold in your head: does
this text read as this language — yes, no, or cannot say. The dependency runs one
way, retrying onto detection, and the call sites already wanted the halves
separately: the audit tools import the check, the simplification services import
the loop.

The tests split the same way, for the same reason.
